(#58) refactor: コンポーネントを切り出した
