+++ frontend/src/app/dashboard/page.tsx
@@ -1,5 +1,7 @@
 "use client"
 import React, { useEffect, useState } from 'react'
+import Sidebar from '../../components/dashboard/Sidebar'
+import AlertItem from '../../components/common/AlertItem'
 
 export default function Page() {
 	const toggleDark = () => {
@@ -9,7 +11,7 @@ export default function Page() {
 	useEffect(() => {
 		if (typeof document === 'undefined') return
 
-		// Load Noto Sans JP if not already present
+		// フォントを読み込む
 		if (!document.querySelector('link[data-dash-font="noto-jp"]')) {
 			const l = document.createElement('link')
 			l.setAttribute('rel', 'stylesheet')
@@ -18,7 +20,7 @@ export default function Page() {
 			document.head.appendChild(l)
 		}
 
-		// Ensure Material Icons are available
+		// Material Iconを読み込む
 		if (!document.querySelector('link[data-dash-icons="material"]')) {
 			const l2 = document.createElement('link')
 			l2.setAttribute('rel', 'stylesheet')
@@ -33,55 +35,8 @@ export default function Page() {
 	return (
 		<div className="min-h-screen bg-background-light dark:bg-background-dark text-gray-800 dark:text-gray-100 font-sans transition-colors duration-200">
 			<div className="flex h-screen overflow-hidden">
-				<aside className={`${sidebarCollapsed ? 'w-20' : 'w-64'} bg-surface-light dark:bg-surface-dark border-r border-border-light dark:border-border-dark flex-col justify-between hidden md:flex transition-all duration-200 z-20 shadow-sm`}>
-					<div>
-						<div className="h-16 flex items-center justify-center border-b border-border-light dark:border-border-dark">
-							<h1 className="text-xl font-bold tracking-wider text-primary dark:text-white uppercase">{sidebarCollapsed ? '' : 'Omamori'}</h1>
-						</div>
-						<nav className="mt-6 px-3 space-y-1">
-							<a className={`relative group flex items-center ${sidebarCollapsed ? 'justify-center px-0 py-3' : 'px-4 py-3'} text-sm font-bold rounded-lg transition-all bg-transparent text-gray-700 hover:bg-warm-orange hover:text-white`} href="#">
-								<span className={`${sidebarCollapsed ? '' : 'mr-3'} material-icons-outlined text-xl`}>dashboard</span>
-								{!sidebarCollapsed && <span className="font-semibold">ダッシュボード</span>}
-							</a>
-							<a className={`group flex items-center ${sidebarCollapsed ? 'justify-center px-0 py-3' : 'px-4 py-3'} text-sm font-medium rounded-lg transition-all bg-transparent text-gray-700 hover:bg-warm-orange hover:text-white`} href="#">
-								<span className={`${sidebarCollapsed ? '' : 'mr-3'} material-icons-outlined text-xl`}>people</span>
-								{!sidebarCollapsed && <>メンバー</>}
-							</a>
-							<a className={`group flex items-center ${sidebarCollapsed ? 'justify-center px-0 py-3' : 'px-4 py-3'} text-sm font-medium rounded-lg transition-all bg-transparent text-gray-700 hover:bg-warm-orange hover:text-white`} href="#">
-								<span className={`${sidebarCollapsed ? '' : 'mr-3'} material-icons-outlined text-xl`}>article</span>
-								{!sidebarCollapsed && <>作業ログ</>}
-							</a>
-							<a className={`group flex items-center ${sidebarCollapsed ? 'justify-center px-0 py-3' : 'px-4 py-3'} text-sm font-medium rounded-lg transition-all bg-transparent text-gray-700 hover:bg-warm-orange hover:text-white`} href="#">
-								<div className="flex items-center">
-									<span className={`${sidebarCollapsed ? '' : 'mr-3'} material-icons-outlined text-xl`}>notifications</span>
-									{!sidebarCollapsed && <>アラート</>}
-								</div>
-								{!sidebarCollapsed && <span className="inline-flex items-center justify-center px-2 py-0.5 ml-3 text-xs font-medium text-white bg-danger rounded-full">3</span>}
-							</a>
-							{/* collapse toggle inserted inline so it appears under Alerts */}
-							<div className={`flex ${sidebarCollapsed ? 'justify-center' : ''} ${sidebarCollapsed ? 'px-2 py-2' : 'px-4 py-2'}`}>
-								<button
-									onClick={() => setSidebarCollapsed(s => !s)}
-									aria-label="toggle sidebar"
-									aria-expanded={!sidebarCollapsed}
-									title={sidebarCollapsed ? 'サイドバーを展開' : 'サイドバーを折りたたむ'}
-									className={sidebarCollapsed
-										? 'w-10 h-10 flex items-center justify-center rounded-lg bg-white shadow-sm border border-border-light text-gray-700 hover:bg-gray-50 transition'
-										: 'w-full flex items-center justify-between px-4 py-2 rounded-lg bg-white border border-border-light text-sm font-medium text-gray-600 hover:bg-gray-100 transition'}
-								>
-									<span className="material-icons-outlined">{sidebarCollapsed ? 'chevron_right' : 'chevron_left'}</span>
-									{!sidebarCollapsed && <span className="ml-2"></span>}
-								</button>
-							</div>
-						</nav>
-					</div>
-					<div className="p-3 border-t border-border-light dark:border-border-dark">
-						<a className={`group flex items-center ${sidebarCollapsed ? 'justify-center px-0 py-3' : 'px-4 py-3'} text-sm font-medium text-gray-600 dark:text-gray-400 hover:bg-gray-100 dark:hover:bg-gray-700 hover:text-gray-900 dark:hover:text-white rounded-lg transition-all`} href="#">
-							<span className={`${sidebarCollapsed ? '' : 'mr-3'} material-icons-outlined text-xl`}>settings</span>
-							{!sidebarCollapsed && <>設定</>}
-						</a>
-					</div>
-				</aside>
+			    {/* サイドバー */}
+					<Sidebar sidebarCollapsed={sidebarCollapsed} setSidebarCollapsed={setSidebarCollapsed} />
 
 				<div className="flex-1 flex flex-col min-w-0 overflow-hidden">
 					<header className="h-16 bg-surface-light dark:bg-surface-dark border-b border-border-light dark:border-border-dark flex items-center justify-between px-4 sm:px-6 lg:px-8 z-10">
@@ -90,6 +45,7 @@ export default function Page() {
 						</button>
 						<div className="flex-1 flex justify-center lg:justify-start lg:ml-6">
 								<div className="w-full max-w-lg lg:max-w-xs relative">
+									{/* 検索バー */}
 									<div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
 										<span className="material-icons-outlined text-warm-brown-600 text-lg">search</span>
 									</div>
@@ -101,6 +57,7 @@ export default function Page() {
 
 					<main className="flex-1 overflow-y-auto focus:outline-none p-4 sm:p-6 lg:p-8">
 						<div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6 mb-8">
+							{/* 作業中メンバー（３つの中の左） */}
 							<div className="bg-white overflow-hidden shadow-sm rounded-xl border border-border-light dark:border-border-dark flex flex-col p-6">
 								<div className="flex items-center justify-between mb-4">
 									<h3 className="text-sm font-medium text-gray-500 dark:text-gray-400">稼働中の作業員</h3>
@@ -115,6 +72,7 @@ export default function Page() {
 								</div>
 							</div>
 
+							{/* 未対応アラート（３つの中の中央） */}
 							<div className="relative bg-white overflow-hidden shadow-lg rounded-xl flex flex-col p-6 border border-border-light dark:border-border-dark">
 								<span className="absolute left-0 top-0 bottom-0 w-1.5 bg-danger rounded-r-md" aria-hidden="true" />
 								<div className="flex items-center justify-between mb-4 pl-4">
@@ -130,6 +88,7 @@ export default function Page() {
 								</div>
 							</div>
 
+							{/* 稼働中の現場（３つの中の右） */}
 							<div className="bg-white overflow-hidden shadow-sm rounded-xl border border-border-light dark:border-border-dark flex flex-col p-6">
 								<div className="flex items-center justify-between mb-4">
 									<h3 className="text-sm font-medium text-gray-500 dark:text-gray-400">稼働中の現場</h3>
@@ -142,7 +101,9 @@ export default function Page() {
 							</div>
 						</div>
 
+						{/* リアルタイムマップと最近のアラート */}
 						<div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
+							{/* リアルタイムマップ（中央） */}
 							<div className="lg:col-span-2 bg-white shadow-sm rounded-xl border border-border-light dark:border-border-dark flex flex-col h-full min-h-[400px]">
 										<div className="p-5 border-b border-border-light dark:border-border-dark flex justify-between items-center">
 											<h2 className="text-lg font-bold text-gray-900 dark:text-white flex items-center"><span className="material-icons-outlined mr-2">map</span>リアルタイムマップ</h2>
@@ -169,16 +130,17 @@ export default function Page() {
 										</div>
 							</div>
 
+              {/* 最近のアラート（右側） */}
 							<div className="bg-white shadow-sm rounded-xl border border-border-light dark:border-border-dark flex flex-col">
 								<div className="p-5 border-b border-border-light dark:border-border-dark">
 									<h2 className="text-lg font-bold text-gray-900 dark:text-white">最近のアラート</h2>
 								</div>
-								<div className="p-4 space-y-4">
-									<AlertItem name="田中 健太" time="10:05" text="動きが検知されません" variant="default" />
-									<AlertItem name="鈴木 一郎" time="09:42" text="SOS信号" variant="important" />
-									<AlertItem name="佐藤 学" time="09:15" text="バッテリー残量低下" variant="default" />
-									<AlertItem name="高橋 エリ" time="08:30" text="エリア入室確認" variant="muted" />
-								</div>
+																<div className="p-4 space-y-4">
+																	<AlertItem name="田中 健太" time="10:05" text="動きが検知されません" variant="default" />
+																	<AlertItem name="鈴木 一郎" time="09:42" text="SOS信号" variant="important" />
+																	<AlertItem name="佐藤 学" time="09:15" text="バッテリー残量低下" variant="default" />
+																	<AlertItem name="高橋 エリ" time="08:30" text="エリア入室確認" variant="muted" />
+																</div>
 								<div className="p-4 border-t border-border-light dark:border-border-dark mt-auto">
 									<button className="w-full py-2 text-sm text-primary dark:text-gray-300 font-medium hover:underline">すべてのアラートを見る</button>
 								</div>
@@ -195,33 +157,4 @@ export default function Page() {
 			</div>
 		</div>
 	)
-}
-
-function AlertItem({ name, time, text, variant = 'default' }: { name: string; time: string; text: string; variant?: 'default' | 'important' | 'muted' }) {
-		const base = 'flex rounded-lg overflow-hidden border';
-		const classes = variant === 'important'
-			? `${base} bg-danger/10 border border-danger/30`
-			: variant === 'muted'
-				? `${base} bg-gray-50 dark:bg-slate-800/50 opacity-60 border border-gray-100 dark:border-gray-700`
-				: `${base} bg-gray-50 dark:bg-slate-800/50 border border-gray-100 dark:border-gray-700`;
-
-		const bar = variant === 'important' ? 'bg-danger' : variant === 'muted' ? 'bg-green-500' : 'bg-gray-400';
-
-		const initials = name.split(/\s+/).map(s => s[0]).filter(Boolean).slice(0,2).join('').toUpperCase()
-
-		return (
-			<div className={classes}>
-				<div className={`w-1.5 ${bar}`}></div>
-				<div className="p-3 w-full">
-					<div className="flex justify-between items-start mb-1">
-						<div className="flex items-center">
-							<div className="w-8 h-8 rounded-full bg-gray-100 dark:bg-gray-700 flex items-center justify-center text-xs font-semibold mr-3 text-gray-700 dark:text-gray-100">{initials}</div>
-							<h4 className="text-sm font-bold text-gray-900 dark:text-white">{name}</h4>
-						</div>
-						<span className="text-xs font-mono text-gray-500 dark:text-gray-400">{time}</span>
-					</div>
-					<p className={`text-sm ${variant === 'important' ? 'text-danger dark:text-red-200 font-medium' : 'text-gray-600 dark:text-gray-300'}`}>{text}</p>
-				</div>
-			</div>
-		)
 }
